backend/requirements.txt
# Azure Functions Core
azure-functions==1.18.0

# Essential Python packages
setuptools>=65.0.0

# Azure Services
azure-cosmos==4.5.1
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0

# Authentication & Security
PyJWT[crypto]==2.8.0
cryptography==41.0.7

# AI/ML Libraries (Essential only)
google-generativeai==0.7.2
numpy>=1.21.0  # Required for vector similarity calculations (lightweight)
# Note: Using Gemini API for embeddings instead of sentence-transformers

# Content Sourcing Dependencies
aiohttp==3.9.1
beautifulsoup4==4.12.2
PyPDF2==3.0.1
lxml==4.9.3
# This saves ~500MB in deployment size and improves cold start times

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
typing-extensions==4.8.0

# Logging and Monitoring
structlog==23.2.0

# Text Processing (lightweight)
beautifulsoup4==4.12.2

# Testing (for CI)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
